code for chat bot/requirements.txt
streamlit
openai>=1.0.0
requests
beautifulsoup4
pillow